🚚 rename path to dms-list
